requirements.txt
arabic-reshaper==3.0.0
asgiref==3.8.1
asn1crypto==1.5.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
cssselect2==0.7.0
daphne==4.1.2
defusedxml==0.8.0rc2
dj-rest-auth==6.0.0
Django==5.1
django-allauth==0.61.1
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangochannelsrestframework==1.2.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
html5lib==1.1
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
lxml==5.3.0
msgpack==1.1.0
oauthlib==3.2.2
oscrypto==1.3.0
pillow==11.0.0
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pyHanko==0.25.1
pyhanko-certvalidator==0.26.3
PyJWT==2.9.0
pyOpenSSL==24.3.0
pypdf==5.1.0
python-barcode==0.15.1
python-bidi==0.6.3
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
qrcode==8.0
redis==5.2.0
reportlab==4.2.5
requests==2.32.3
requests-oauthlib==2.0.0
service-identity==24.2.0
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
svglib==1.5.1
tinycss2==1.4.0
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
uritools==4.0.3
urllib3==2.2.3
webencodings==0.5.1
xhtml2pdf==0.2.16
zope.interface==7.2
